Updated cxxtest to version 4.4. The new version of cxxtest works with latest versions of python and will hopefully fix the intermittent build failures.

+++ Source/RayLibs/TestTrace/TestTracer.h
@@ -174,7 +174,7 @@ inline void TestTracer::RenderImage()
   intrusive_ptr<PhotonLTEIntegrator> p_lte_int( new PhotonLTEIntegrator(mp_scene, params) );
 
   t0 = tbb::tick_count::now();
-  p_lte_int->ShootPhotons(50*1000000U, true);
+  p_lte_int->ShootPhotons(50*(size_t)1000000, true);
   t1 = tbb::tick_count::now();
   printf("Shooting: %lf\n", (t1-t0).seconds());
